# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/habits": {
      "post": {
        "operationId": "HabitController_createHabit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHabitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "习惯创建成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建新习惯",
        "tags": [
          "习惯打卡"
        ]
      },
      "get": {
        "operationId": "HabitController_getUserHabits",
        "parameters": [
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "description": "是否包含已归档的习惯",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取用户的所有习惯",
        "tags": [
          "习惯打卡"
        ]
      }
    },
    "/habits/{habitId}": {
      "get": {
        "operationId": "HabitController_getHabitById",
        "parameters": [
          {
            "name": "habitId",
            "required": true,
            "in": "path",
            "description": "习惯ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "年月，格式为YYYY-MM",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "习惯不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取单个习惯详情",
        "tags": [
          "习惯打卡"
        ]
      },
      "put": {
        "operationId": "HabitController_updateHabit",
        "parameters": [
          {
            "name": "habitId",
            "required": true,
            "in": "path",
            "description": "习惯ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHabitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "习惯不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新习惯",
        "tags": [
          "习惯打卡"
        ]
      },
      "delete": {
        "operationId": "HabitController_deleteHabit",
        "parameters": [
          {
            "name": "habitId",
            "required": true,
            "in": "path",
            "description": "习惯ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "习惯不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除习惯（软删除）",
        "tags": [
          "习惯打卡"
        ]
      }
    },
    "/habits/{habitId}/check-ins": {
      "post": {
        "operationId": "HabitController_createCheckIn",
        "parameters": [
          {
            "name": "habitId",
            "required": true,
            "in": "path",
            "description": "习惯ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHabitCheckInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "打卡成功"
          },
          "400": {
            "description": "已打卡或参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "习惯不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建打卡记录",
        "tags": [
          "习惯打卡"
        ]
      },
      "get": {
        "operationId": "HabitController_getCheckIns",
        "parameters": [
          {
            "name": "habitId",
            "required": true,
            "in": "path",
            "description": "习惯ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "开始日期",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "结束日期",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "习惯不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取习惯的打卡记录",
        "tags": [
          "习惯打卡"
        ]
      }
    },
    "/habits/{habitId}/toggle-check-in": {
      "post": {
        "operationId": "HabitController_toggleCheckInStatus",
        "parameters": [
          {
            "name": "habitId",
            "required": true,
            "in": "path",
            "description": "习惯ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "状态切换成功"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "习惯不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "切换打卡状态",
        "tags": [
          "习惯打卡"
        ]
      }
    },
    "/habits/{habitId}/streak": {
      "get": {
        "operationId": "HabitController_getHabitStreak",
        "parameters": [
          {
            "name": "habitId",
            "required": true,
            "in": "path",
            "description": "习惯ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "习惯不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取习惯的连续打卡统计",
        "tags": [
          "习惯打卡"
        ]
      }
    }
  },
  "info": {
    "title": "叮咚待办 (DingDongTodo) API",
    "description": "DingDongTodo后端API文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "JWT认证Bearer Token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateHabitDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateHabitDto": {
        "type": "object",
        "properties": {}
      },
      "CreateHabitCheckInDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}